refacto(parsing): renamed to `read`
